Change to class-based to function-based
