+++ .flowconfig
@@ -1,2 +1,5 @@
 [ignore]
 .*/node_modules/.*
+
+[libs]
+./flow-typed
