+++ flow-typed/manual-installed-npm/object-path-immutable_vx.x.x.js
@@ -1,0 +1,3 @@
+declare module 'object-path-immutable' {
+  declare function set<O>(obj: O, path: (string | (number | string)[]), value: any): O;
+}
