+++ lib/border-utils.js
@@ -1,7 +1,18 @@
 // @flow
 
+const matrixUtils = require('./matrix-utils');
+
 /*::
-import type Box, {ElementSymbol, Rectangle} from './Box';
+import type {
+  Element,
+  ElementSymbol,
+  Matrix,
+} from './matrix-utils';
+import type {Rectangle} from './rectangle-utils';
+
+import type {
+  Box,
+} from './box-utils';
  */
 
 
@@ -9,36 +20,54 @@ const getByCirculatedIndex = (ary/*: any[]*/, index/*: number*/)/*: any*/ => {
   return ary[(ary.length + index) % ary.length];
 };
 
-const clearTopSide = (box/*: Box*/, width/*: number*/)/*: void*/ => {
-  for (let y = 0; y < width; y += 1) {
-    const row = box.getMatrix()[y];
-    row.forEach(element => { element.symbol = null; });
-  }
+const clearTopSide = (box/*: Box*/, width/*: number*/)/*: Box*/ => {
+  const newMatrix = box.matrix.map((row, y) => {
+    if (y < width) {
+      return row.map(element => Object.assign({}, element, {symbol: null}));
+    }
+    return row;
+  });
+
+  return Object.assign({}, box, {matrix: newMatrix});
 };
 
-const clearBottomSide = (box/*: Box*/, width/*: number*/)/*: void*/ => {
-  const maxY = box.getMaxY();
-  for (let y = maxY; y > maxY - width; y -= 1) {
-    const row = box.getMatrix()[y];
-    row.forEach(element => { element.symbol = null; });
-  }
+const clearBottomSide = (box/*: Box*/, width/*: number*/)/*: Box*/ => {
+  const maxY = matrixUtils.getMaxY(box.matrix);
+
+  const newMatrix = box.matrix.map((row, y) => {
+    if (y > maxY - width) {
+      return row.map(element => Object.assign({}, element, {symbol: null}));
+    }
+    return row;
+  });
+
+  return Object.assign({}, box, {matrix: newMatrix});
 };
 
-const clearLeftSide = (box/*: Box*/, width/*: number*/)/*: void*/ => {
-  box.getMatrix().forEach(row => {
+const clearLeftSide = (box/*: Box*/, width/*: number*/)/*: Box*/ => {
+  const newMatrix = box.matrix.map(row => {
+    const newRow = row.slice();
     for (let x = 0; x < width; x += 1) {
-      row[x].symbol = null;
+      newRow[x] = Object.assign({}, newRow[x], {symbol: null});
     }
+    return newRow;
   });
+
+  return Object.assign({}, box, {matrix: newMatrix});
 };
 
-const clearRightSide = (box/*: Box*/, width/*: number*/)/*: void*/ => {
-  const maxX = box.getMaxX();
-  box.getMatrix().forEach(row => {
+const clearRightSide = (box/*: Box*/, width/*: number*/)/*: Box*/ => {
+  const maxX = matrixUtils.getMaxX(box.matrix);
+
+  const newMatrix = box.matrix.map(row => {
+    const newRow = row.slice();
     for (let x = maxX; x > maxX - width; x -= 1) {
-      row[x].symbol = null;
+      newRow[x] = Object.assign({}, newRow[x], {symbol: null});
     }
+    return newRow;
   });
+
+  return Object.assign({}, box, {matrix: newMatrix});
 };
 
 const drawTopSide = (
@@ -47,15 +76,22 @@ const drawTopSide = (
   symbols/*: (ElementSymbol | null)[]*/,
   fromX/*: number*/,
   toX/*: number*/
-)/*: void*/ => {
-  for (let x = fromX; x <= toX; x += 1) {
-    for (let y = 0; y < borderWidth; y += 1) {
-      const element = box.getElement({x, y});
-      if (element && symbols.length > 0) {
-        element.symbol = getByCirculatedIndex(symbols, y);
-      }
+)/*: Box*/ => {
+  const newMatrix = box.matrix.map((row, y) => {
+    if (y < borderWidth) {
+      return row.map((element, x) => {
+        if (x >= fromX && x <= toX && symbols.length > 0) {
+          return Object.assign({}, element, {
+            symbol: getByCirculatedIndex(symbols, y),
+          });
+        }
+        return element;
+      });
     }
-  }
+    return row;
+  });
+
+  return Object.assign({}, box, {matrix: newMatrix});
 };
 
 const drawBottomSide = (
@@ -64,17 +100,24 @@ const drawBottomSide = (
   symbols/*: (ElementSymbol | null)[]*/,
   fromX/*: number*/,
   toX/*: number*/
-)/*: void*/ => {
-  const maxY = box.getMaxY();
-  for (let x = fromX; x <= toX; x += 1) {
-    for (let deltaY = 0; deltaY < borderWidth; deltaY += 1) {
-      const y = maxY - deltaY;
-      const element = box.getElement({x, y});
-      if (element && symbols.length > 0) {
-        element.symbol = getByCirculatedIndex(symbols, deltaY);
-      }
+)/*: Box*/ => {
+  const maxY = matrixUtils.getMaxY(box.matrix);
+
+  const newMatrix = box.matrix.map((row, y) => {
+    if (y > maxY - borderWidth) {
+      return row.map((element, x) => {
+        if (x >= fromX && x <= toX && symbols.length > 0) {
+          return Object.assign({}, element, {
+            symbol: getByCirculatedIndex(symbols, maxY - y),
+          });
+        }
+        return element;
+      });
     }
-  }
+    return row;
+  });
+
+  return Object.assign({}, box, {matrix: newMatrix});
 };
 
 const drawLeftSide = (
@@ -83,15 +126,22 @@ const drawLeftSide = (
   symbols/*: (ElementSymbol | null)[]*/,
   fromY/*: number*/,
   toY/*: number*/
-)/*: void*/ => {
-  for (let y = fromY; y <= toY; y += 1) {
-    for (let x = 0; x < borderWidth; x += 1) {
-      const element = box.getElement({x, y});
-      if (element && symbols.length > 0) {
-        element.symbol = getByCirculatedIndex(symbols, x);
-      }
+)/*: Box*/ => {
+  const newMatrix = box.matrix.map((row, y) => {
+    if (y >= fromY && y <= toY) {
+      return row.map((element, x) => {
+        if (x < borderWidth && symbols.length > 0) {
+          return Object.assign({}, element, {
+            symbol: getByCirculatedIndex(symbols, x),
+          });
+        }
+        return element;
+      });
     }
-  }
+    return row;
+  });
+
+  return Object.assign({}, box, {matrix: newMatrix});
 };
 
 const drawRightSide = (
@@ -100,36 +150,50 @@ const drawRightSide = (
   symbols/*: (ElementSymbol | null)[]*/,
   fromY/*: number*/,
   toY/*: number*/
-)/*: void*/ => {
-  const maxX = box.getMaxX();
-  for (let y = fromY; y <= toY; y += 1) {
-    for (let deltaX = 0; deltaX < borderWidth; deltaX += 1) {
-      const x = maxX - deltaX;
-      const element = box.getElement({x, y});
-      if (element && symbols.length > 0) {
-        element.symbol = getByCirculatedIndex(symbols, deltaX);
-      }
+)/*: Box*/ => {
+  const maxX = matrixUtils.getMaxX(box.matrix);
+
+  const newMatrix = box.matrix.map((row, y) => {
+    if (y >= fromY && y <= toY) {
+      return row.map((element, x) => {
+        if (x > maxX - borderWidth && symbols.length > 0) {
+          return Object.assign({}, element, {
+            symbol: getByCirculatedIndex(symbols, maxX - x),
+          });
+        }
+        return element;
+      });
     }
-  }
+    return row;
+  });
+
+  return Object.assign({}, box, {matrix: newMatrix});
 };
 
 const drawCorner = (
   box/*: Box*/,
   rectangle/*: Rectangle*/,
   symbols/*: (ElementSymbol | null)[]*/
-)/*: void*/ => {
+)/*: Box*/ => {
   let counter = 0;
-  for (let y = rectangle.y; y < rectangle.y + rectangle.height; y += 1) {
-    for (let x = rectangle.x; x < rectangle.x + rectangle.width; x += 1) {
-      const element = box.getElement({x, y});
-      if (element) {
-        element.symbol = getByCirculatedIndex(symbols, counter);
-      } else {
-        throw new Error('The coordinate is invalid');
-      }
-      counter += 1;
+
+  const newMatrix = box.matrix.map((row, y) => {
+    if (y >= rectangle.y && y < rectangle.y + rectangle.height) {
+      return row.map((element, x) => {
+        if (x >= rectangle.x && x < rectangle.x + rectangle.width) {
+          const newRow = Object.assign({}, element, {
+            symbol: getByCirculatedIndex(symbols, counter),
+          });
+          counter += 1;
+          return newRow;
+        }
+        return element;
+      });
     }
-  }
+    return row;
+  });
+
+  return Object.assign({}, box, {matrix: newMatrix});
 };
 
 module.exports = {
