+++ lib/coordinate-utils.js
@@ -1,0 +1,6 @@
+/*::
+export type Coordinate = {
+  x: number,
+  y: number,
+};
+ */
