+++ package.json
@@ -22,6 +22,7 @@
     "url": "https://github.com/kjirou/text-sketchbook/issues"
   },
   "homepage": "https://github.com/kjirou/text-sketchbook#readme",
+  "dependencies": {},
   "devDependencies": {
     "flow-bin": "^0.61.0",
     "mocha": "^3.5.0"
